enhanced styles. moved proposal to Lock YFD Menu
