+++ src/assets/styles/scss/_root.module.scss
@@ -14,6 +14,7 @@
     --theme-bg-menu: rgba(32, 36, 46, 0.8);
     --theme-bg-filter: drop-shadow(0 0 10px var(--theme-gray-dark));
     --theme-bg-menu-hover: rgba(254, 254, 255, 0.3);
+    --theme-bg-icon-hover: var(--theme-inactive-color);
     
     /* colors */
     --theme-color: var(--theme-text);
