+++ src/assets/styles/scss/app.module.scss
@@ -9,7 +9,7 @@
 
 @import "_root.module.scss";
 @import "_image.module.scss";
-@import "color.module.scss";
+@import "icon.module.scss";
 @import "nav/main.module.scss";
 @import "wrapper.module.scss";
 @import "header.module.scss";
