+++ src/assets/styles/scss/side.module.scss
@@ -66,4 +66,8 @@
         @extend i;
         height: 1em;
     }
+
+    span {
+        margin-right: 0.5rem;
+    }
 }
