+++ src/features/forms/feedback/ModalButton.tsx
@@ -1,6 +1,6 @@
 import { Text } from '@chakra-ui/react';
 
-import styles from '@scss/color.module.scss';
+import styles from '@scss/icon.module.scss';
 import { Icons } from '@var/icons';
 
 function FeedbackModalButton({ onOpen }: any) {
