+++ src/features/nav/left/initiatives.tsx
@@ -4,9 +4,6 @@ import { useDisclosure } from '@chakra-ui/react';
 import styles from '@scss/side.module.scss';
 import { Icons } from '@var/icons';
 
-import ProposalModal from '@features/proposals/modal';
-import ProposalModalButton from '@features/proposal/proposalCreationModal/ProposalModalButton';
-
 export default function MenuLeftInitiatives() {
   const { isOpen, onOpen, onClose } = useDisclosure();
 
@@ -15,8 +12,6 @@ export default function MenuLeftInitiatives() {
       <div className={styles['side-wrapper-menu']}>
         <div className={styles['side-title']}>Initatives</div>
         <div className={styles['side-menu']}>
-          <ProposalModalButton onOpen={onOpen} />
-          <ProposalModal isOpen={isOpen} onClose={onClose} />
           <Link to="/proposals-vaults">
             <i className="material-symbols-outlined">{Icons.vault}</i>
             Vaults
